+++ src/main/java/com/service/usermanagement/repositories/UserRepository.java
@@ -5,7 +5,7 @@ import org.apache.ibatis.annotations.Param;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface UserRepository extends JpaRepository<UserEntity, Integer> {
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
 
     @Query("SELECT u FROM UserEntity u WHERE u.email=:email")
     UserEntity findByEmail(@Param("email") String email);
